+++ lib/fl_border_gradient.dart
@@ -1,6 +1,6 @@
-library fl_border_gradient;
-
-export 'src/gradient_box_border.dart';
 export 'src/animated_gradient_border.dart';
+export 'src/fl_border_gradient_theme.dart';
 export 'src/glowing_gradient_border.dart';
+export 'src/gradient_box_border.dart';
+export 'src/gradient_sides_box_border.dart';